Relatório Fatorial Adicionado

+++ AED2/FATORIAL/Fatorial.c
@@ -1,6 +1,6 @@
 #include<stdio.h>
 
-float Fatorial(int);
+double Fatorial(int);
 
 
 int recur = 0;
@@ -8,13 +8,13 @@ int main(){
   for(int i = 10; i<=100;i+=10){
     recur = 0;
     printf("Fatorial de %d\n",i);
-    printf("eh %.0f\n",Fatorial(i));
+    printf("eh %.1lf\n",Fatorial(i));
     printf("total recursões %d\n--------\n",recur);
   }
   return 0;
 }
 
-float Fatorial(int p){
+double Fatorial(int p){
   //printf("recursão %d\n",recur);
   recur++;
   if(p == 0){
